// src/options/app/lib/type.ts
export type OptionID = string
export type OptionUI_PageID = string
export type OptionUI_CategoryID = string
export type OptionUI_ItemID = string
// ...
export class OptionUI_Descriptions {
    attention?: string
    text?: string
    small?: string
    hidden?: string
    keywords?: Array<string>
}
// ...
export type OptionUI_Value_RequirementCompareMethods = "="|"!="|">"|"<"|">="|"<="|"in"|"of"|"include"
export interface OptionUI_Item {
    id: OptionUI_ItemID
    title: string
    description?: OptionUI_Descriptions
    location: {
        page: OptionUI_PageID
        category: OptionUI_CategoryID
        hide?: boolean
        noindex?: boolean
        parent?: string
    }
    ui?: OptionUI_Item_UI
    value?: {
        related?: Array<OptionID> | "children"
        isAdvanced?: boolean
        isDebug?: boolean
        isExperimental?: boolean
        requirement?: {
            condition?: "and"|"or"|"nor"|"nand"
            requirementParams?: Array<{
                id: OptionID,
                value: any,
                /**
                 * `=`: `storage` === `v` (default)  
                 * `!=`: `storage` !== `v`  
                 * `>`: `storage` > `v`  
                 * `<`: `storage` < `v`  
                 * `>=`: `storage` >= `v`  
                 * `<=`: `storage` <= `v`  
                 * `<=`: `storage` <= `v`  
                 * `in`: `v` in `storage`  
                 * `of`: `v`.includes(`storage`)  
                 * `include`: `storage`.includes(`v`)  
                 */
                compare?: OptionUI_Value_RequirementCompareMethods
            }>
        }
    }
}

interface OptionUI_Item_UI<T = "switch" | "checkbox" | "select" | "radio" | "slider" | "textarea" | "code"
| "parent" | "custom" | "color" | "textfield" | "number"> {
    type: T
    variant?: "filled" | "standard" | "outlined"
    data?: T extends "switch" | "checkbox"
        ? OptionUI_Item_Switch
        : T extends "select"
        ? OptionUI_Item_Select
        : T extends "radio"
        ? OptionUI_Item_Select
        : T extends "slider"
        ? OptionUI_Item_Slider
        : T extends "textarea"
        ? OptionUI_Item_TextArea
        : T extends "code"
        ? OptionUI_Item_Code
        : T extends "color"
        ? OptionUI_Item_Color
        : T extends "textfield"
        ? OptionUI_Item_TextField
        : T extends "number"
        ? OptionUI_Item_Number
        : T extends "parent"
        ? OptionUI_Item_Parent
        : T extends "custom"
        ? OptionUI_Item_Custom
        : {}
    label?: string
    description?: string
    placeholder?: string
    labelPlacement?: 'bottom'|'end'|'start'|'top'
    showForce?: boolean
    hideButtons?: Array<"favorite"|"reset"|"info">|"all"
    hideDivider?: boolean
}


/************************************************************************************* */
/* オプションページ データ */
export interface OptionUI_ItemProps {
    option: OptionUI_Item
    storage?: nt.storage.local.options
    type?: "search" | "favorite" | "default"
    child?: number
}
// ...
export class OptionUI_Item_Radio<U = "string"|"number"> {
    dataType?: U
    values?: U extends "string"
        ? Array<{ value: string, label: string }>
        : U extends "number"
        ? Array<{ value: number, label: string }>
        : Array<{ value: string|number, label: string }>
    layout?: "default"|"wide"
}